frontend/requirements.txt
# Streamlit
streamlit==1.28.2

# Data visualization
plotly==5.18.0
altair==5.1.2
pandas==2.1.3

# API client
httpx==0.26.0
requests==2.31.0

# Authentication
supabase==2.9.0
websockets==13.0

# Utilities
python-dotenv==1.0.0
